# data/JAMII_.swagger_collection.yml
openapi: 3.0.0
info:
  title: JAMII_MATERNITY_HOSPITAL
  version: 1.0.0
servers:
  - url: https://jamii-maternity-hospital.vercel.app
components:
  securitySchemes:
    noauthAuth:
      type: http
      scheme: noauth
    apikeyAuth:
      type: http
      scheme: apikey
...
paths:
...
  /api/surfermessages:
    post:
      tags:
        - Surfer Messages
      summary: ✅ Create Surfermessage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                firstName: Last
                lastName: Sicko
                email: last.sicko@gmail.com
                surferMessage: Healed, come for a treat nurse User...
      security:
        - apikeyAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - Surfer Messages
      summary: ✅ Get All Surfermessage (FB)
      security:
        - apikeyAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...